backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
openai==1.58.1
google-generativeai==0.8.3
anthropic==0.40.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-multipart==0.0.17
fastapi-cors==0.0.6
diskcache==5.6.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-requests
